Deeply merge yamls in original data
so that blocks can be defined in more than one yaml files.

This will allow us to write down the
`uuid_label_mappings.COS_PERSISTENT` key in a new file if it's not
already in the configuration, without having to worry in which file
`kcrypt.challenger_server` has been set.

There are 2 different keys in the same block (`kcrypt`) and the previous
code didn't allow blocks to be defined in multiple files.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>